requirements.txt
pypdf 
dill
ipdb
pyperclip
flask
werkzeug
openai==0.28
python-dotenv
hypercorn
reportlab
python-docx
